[z0187] Store some config values in type specific emuns/arrays.
(based on master commit 7fff320)
(based on master commit ac1ce74)
(based on master commit 050552a)
(based on master commit fc68c44)
(based on master commit bc26ed8)

(based on commit 3add076)

+++ src/game/GMTicketMgr.cpp
@@ -47,7 +47,7 @@ void GMTicketMgr::LoadGMTickets()
         return;
     }
 
-    barGoLink bar( result->GetRowCount() );
+    barGoLink bar( (int)result->GetRowCount() );
 
     uint32 count = 0;
